+++ packages/bridge-status-controller/src/bridge-status-controller.ts
@@ -1333,6 +1333,8 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       activeAbTests?: { key: string; value: string }[];
       tokenSecurityTypeDestination?: string | null;
       inputPrimaryDenomination?: InputPrimaryDenomination;
+      customSlippage?: boolean;
+      slippagePercentage?: number;
     },
   ): Promise<TransactionMeta> => {
     let tradeTxMeta!: TransactionMeta;
@@ -1379,7 +1381,10 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
               quoteResponse: payload.quoteResponse,
               accountAddress: params.selectedAccount.address,
               isStxEnabled: params.isStxEnabled,
-              slippagePercentage: payload.quoteResponse.quote.slippage ?? 0,
+              slippagePercentage:
+                sharedHistoryItemProperties.slippagePercentage ??
+                payload.quoteResponse.quote.slippage ??
+                0,
             });
             this.#reportSubmittedForNonEvmTx(
               payload.historyKey,
@@ -1516,6 +1521,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       tokenSecurityTypeDestination,
       batchSellTrades,
       batchId,
+      quotesReceivedContext,
     );
 
     try {
@@ -1571,6 +1577,8 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
             activeAbTests,
             tokenSecurityTypeDestination,
             inputPrimaryDenomination,
+            customSlippage: quotesReceivedContext?.custom_slippage,
+            slippagePercentage: quotesReceivedContext?.slippage_limit,
           }),
       );
     } catch (error) {
@@ -1626,7 +1634,6 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       quotesReceivedContext,
     } = params;
 
-    // TODO add metrics context
     return await this.submitTx(
       accountAddress,
       quoteResponse,